swap: move consts into swap/api.go, move api tests into api_test.go
